flask_app/requirements.txt
flask>=3.0.0
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.1
flask-cors>=4.0.0
firebase-admin>=6.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.0
